requirements.txt
pytest==7.1.2
web3==5.30.0
requests==2.28.1
tweepy==4.10.0
python-telegram-bot==13.13
python-dotenv==0.20.0